Consolidate algorithm normalization
This commit splits the Algorithm Normalization algorithm away from the
algorithm.go file into a dedicated normalization.go file. The reasoning
behind that was that we do all sorts of normalization and validation in
the context of the project, and it would be more thematic and easier to
maintain/comprehend if those functions and types lived in a dedicated
file.
