+++ webcrypto/algorithm.go
@@ -4,7 +4,6 @@ import (
 	"crypto"
 	"fmt"
 	"hash"
-	"reflect"
 	"strings"
 
 	"github.com/dop251/goja"
@@ -19,6 +18,49 @@ type Algorithm struct {
 // It implements the NormalizedIdentifier interface.
 func (a Algorithm) NormalizedName() AlgorithmIdentifier {
 	return a.Name
+}
+
+// NewAlgorithm creates a new Algorithm instance from a goja.Value.
+func NewAlgorithm(rt *goja.Runtime, v goja.Value) (Algorithm, error) {
+	if v == nil {
+		return Algorithm{}, NewError(0, SyntaxError, "algorithm is required")
+	}
+
+	var params Algorithm
+	if err := rt.ExportTo(v, &params); err != nil {
+		return Algorithm{}, NewError(0, SyntaxError, "algorithm is invalid")
+	}
+
+	if err := params.Validate(); err != nil {
+		return Algorithm{}, err
+	}
+
+	if err := params.Normalize(); err != nil {
+		return Algorithm{}, err
+	}
+
+	return params, nil
+}
+
+// Validate validates the Algorithm instance fits the specifications
+// requirements. It implements the Validator interface.
+func (a Algorithm) Validate() error {
+	if a.Name == "" {
+		return NewError(0, SyntaxError, "name property is required")
+	}
+
+	if !IsAlgorithm(a.Name) && !IsHashAlgorithm(a.Name) {
+		return NewError(0, NotSupportedError, "algorithm name is not supported")
+	}
+
+	return nil
+}
+
+// Normalize normalizes the Algorithm instance. It implements the Normalizer
+// interface.
+func (a *Algorithm) Normalize() error {
+	a.Name = NormalizeAlgorithmName(a.Name)
+	return nil
 }
 
 // AlgorithmIdentifier represents the name of an algorithm.
@@ -62,42 +104,6 @@ const (
 	ECDH = "ECDH"
 )
 
-// NormalizeAlgorithmName returns the normalized algorithm name.
-//
-// As the algorithm name is case-insensitive, we normalize it to
-// our internal representation.
-func NormalizeAlgorithmName(name string) AlgorithmIdentifier {
-	algorithms := [...]AlgorithmIdentifier{
-		// RSA
-		RSASsaPkcs1v15,
-		RSAPss,
-		RSAOaep,
-
-		// HMAC
-		HMAC,
-
-		// AES
-		AESCtr,
-		AESCbc,
-		AESGcm,
-		AESKw,
-
-		// ECDSA
-		ECDSA,
-
-		// ECDH
-		ECDH,
-	}
-
-	for _, alg := range algorithms {
-		if strings.EqualFold(name, alg) {
-			return alg
-		}
-	}
-
-	return name
-}
-
 // HashAlgorithmIdentifier represents the name of a hash algorithm.
 //
 // Note that it is defined as an alias of string, instead of a dedicated type,
@@ -132,99 +138,6 @@ func Hasher(algorithm HashAlgorithmIdentifier) (func() hash.Hash, error) {
 	}
 
 	return nil, NewError(0, ImplementationError, fmt.Sprintf("unsupported hash algorithm: %s", algorithm))
-}
-
-// NormalizedAlgorithm represents a normalized algorithm.
-type NormalizedAlgorithm interface {
-	NormalizedName() AlgorithmIdentifier
-}
-
-// normalize algorithm
-// normalizeAlgorithm(algorithm: string | Algorithm, op: string):
-
-// NormalizeAlgorithm normalizes the given algorithm following the algorithm described in the WebCrypto [specification].
-//
-// [specification]: https://www.w3.org/TR/WebCryptoAPI/#algorithm-normalization-normalize-an-algorithm
-func NormalizeAlgorithm(rt *goja.Runtime, algorithm goja.Value, op OperationIdentifier) (NormalizedAlgorithm, error) {
-	// 1.
-	registeredAlgorithms, ok := supportedAlgorithms[op]
-	if !ok {
-		return Algorithm{}, NewError(0, ImplementationError, fmt.Sprintf("unsupported operation: %s", op))
-	}
-
-	// 2.
-	var initialAlg Algorithm
-	switch algorithm.ExportType().Kind() {
-	case reflect.String:
-		obj := rt.NewObject()
-		err := obj.Set("name", algorithm.Export())
-		if err != nil {
-			// 3.
-			return Algorithm{}, NewError(0, ImplementationError, "unable to convert the string argument to an object")
-		}
-		return NormalizeAlgorithm(rt, obj, op)
-	case reflect.Map, reflect.Struct:
-		err := rt.ExportTo(algorithm, &initialAlg)
-		if err != nil {
-			// 3.
-			return Algorithm{}, NewError(0, SyntaxError, "algorithm object is not a valid algorithm")
-		}
-	default:
-		return Algorithm{}, NewError(0, SyntaxError, "unsupported algorithm type")
-	}
-
-	// 4.
-	algName := initialAlg.Name
-
-	// 5.
-	var desiredType string
-	algNameRegistered := false
-	for key, value := range registeredAlgorithms {
-		if strings.EqualFold(key, algName) {
-			algName = key
-			desiredType = value
-			algNameRegistered = true
-			break
-		}
-	}
-
-	if !algNameRegistered {
-		return Algorithm{}, NewError(0, NotSupportedError, fmt.Sprintf("unsupported algorithm name: %s", algName))
-	}
-
-	// No further operation is needed if the algorithm does not have a desired type.
-	if desiredType == "" {
-		return Algorithm{Name: algName}, nil
-	}
-
-	// 6.
-	err := NewError(0, ImplementationError, fmt.Sprintf("unsupported algorithm type: %s", desiredType))
-	return Algorithm{}, err
-}
-
-// As defined by the [specification]
-// [specification]: https://w3c.github.io/webcrypto/#algorithm-normalization-internal
-//
-//nolint:gochecknoglobals
-var supportedAlgorithms = map[OperationIdentifier]map[AlgorithmIdentifier]string{
-	OperationIdentifierDigest: {
-		Sha1:   "",
-		Sha256: "",
-		Sha384: "",
-		Sha512: "",
-	},
-	OperationIdentifierGenerateKey: {
-		RSASsaPkcs1v15: "RsaHashedKeyGenParams",
-		RSAPss:         "RsaHashedKeyGenParams",
-		RSAOaep:        "RsaHashedKeyGenParams",
-		ECDSA:          "EcKeyGenParams",
-		ECDH:           "EcKeyGenParams",
-		HMAC:           "HmacKeyGenParams",
-		AESCtr:         "AesKeyGenParams",
-		AESCbc:         "AesKeyGenParams",
-		AESGcm:         "AesKeyGenParams",
-		AESKw:          "AesKeyGenParams",
-	},
 }
 
 // IsAlgorithm returns true if the given algorithm is supported by the library.
@@ -278,47 +191,3 @@ func IsHashAlgorithm(algorithm string) bool {
 
 	return false
 }
-
-// OperationIdentifier represents the name of an operation.
-//
-// Note that it is defined as an alias of string, instead of a dedicated type,
-// to ensure it is handled as a string by goja.
-type OperationIdentifier = string
-
-const (
-	// OperationIdentifierSign represents the sign operation.
-	OperationIdentifierSign OperationIdentifier = "sign"
-
-	// OperationIdentifierVerify represents the verify operation.
-	OperationIdentifierVerify OperationIdentifier = "verify"
-
-	// OperationIdentifierEncrypt represents the encrypt operation.
-	OperationIdentifierEncrypt OperationIdentifier = "encrypt"
-
-	// OperationIdentifierDecrypt represents the decrypt operation.
-	OperationIdentifierDecrypt OperationIdentifier = "decrypt"
-
-	// OperationIdentifierDeriveBits represents the deriveBits operation.
-	OperationIdentifierDeriveBits OperationIdentifier = "deriveBits"
-
-	// OperationIdentifierDeriveKey represents the deriveKey operation.
-	OperationIdentifierDeriveKey OperationIdentifier = "deriveKey"
-
-	// OperationIdentifierWrapKey represents the wrapKey operation.
-	OperationIdentifierWrapKey OperationIdentifier = "wrapKey"
-
-	// OperationIdentifierUnwrapKey represents the unwrapKey operation.
-	OperationIdentifierUnwrapKey OperationIdentifier = "unwrapKey"
-
-	// OperationIdentifierImportKey represents the importKey operation.
-	OperationIdentifierImportKey OperationIdentifier = "importKey"
-
-	// OperationIdentifierExportKey represents the exportKey operation.
-	OperationIdentifierExportKey OperationIdentifier = "exportKey"
-
-	// OperationIdentifierGenerateKey represents the generateKey operation.
-	OperationIdentifierGenerateKey OperationIdentifier = "generateKey"
-
-	// OperationIdentifierDigest represents the digest operation.
-	OperationIdentifierDigest OperationIdentifier = "digest"
-)
